Code reworked to better follow Unity design principles (and work)
Design of first SignalComps changed to better work with Unity, and the prefabs of the Bias and Splitter have been built to reflect these changes.
Splitter, Bias and Port all work as they are supposed to, but are manually set up using unity's inspector
The SignalSource component just holds a value at the moment, in order to act as the start of the chain.
Next job will be to create user input for connecting ports.
After that, a rough timeline will look like:
  -More Components
  -Physical Space and user input to interact with it
  -More complex signalsource component - allowing for more complex varying signals
  -Way of calculating correct input and checking against actual output
  -Structuring to be more like an actual game rather than a tech demo