Port Connection and Prefabs
Ports can be connected by clicking on one, and then another. I don't  believe there is anything stopping outputs connecting to outputs, but this is a quick thing to check and change.
Right click can be used to break a connection and cancel a pending connection.

I also made the first three components into prefabs (finally) with ports linked to the component and set up, but not connected to other signal comps.