Started the source signal coding
The current code can simulate 4 different types of signal.
The script should be attached to a PORT not the SignalSource GameObject. This is so that signal sources can emit multiple signals from multiple ports.
ALSO - as of this commit, I haven't tested it properly. I'm assuming it works because it makes me happy, and I want it committed if Unity crashes.
NEXT - will be to write a way for input signals to be checked against an output signal (and make an output component for it)