SignalInput and SignalOutput has been started
These would be the pieces that hold the information that would allow the SignalOutput to check if the values reaching it are correct in the context of a puzzle.
I haven't done much, and I might start again. I need to think about how the checks will be performed in a compartmentalised way.
It might be that a script will create and instantiate the pairs, hold a ref to both, and check the outputs.
It might be that the pairs just hold different numbers of Inputs/Outputs and use an external script to check them based on which ever "puzzle" has been selected.

I need to think